+++ src/lib.rs
@@ -5,13 +5,13 @@
 //! You need to call [Datadog::init] in your main binary, and to do so you'll need as argument a type that implements the [Configuration] trait.
 //! If you never call [Datadog::init] in your binary NO metrics will be sent.
 //!
-//! Inside the [configuration] you'll find an [implementation of this trait][configuration::Config] tailored for prima.it needs.
+//! Inside the [configuration] you'll find an [implementation of this trait][configuration::Configuration] tailored for prima.it needs.
 //!
 //! ```
-//! use prima_datadog::{*, configuration::Config};
+//! use prima_datadog::{*, configuration::Configuration};
 //!
-//! // initializes the Config struct
-//! let configuration = Config::new(
+//! // initializes the Configuration struct
+//! let configuration = Configuration::new(
 //!     "0.0.0.0:1234", // to address
 //!     "namespace", // namespace for all metrics
 //! );
@@ -28,8 +28,8 @@
 //! - a list of tags (which is separated from the rest of the arguments by semicolon `;`) in the form of `"name" => "value"`
 //!
 //! ```
-//! # use prima_datadog::{*, configuration::Config};
-//! # let configuration = Config::new(
+//! # use prima_datadog::{*, configuration::Configuration};
+//! # let configuration = Configuration::new(
 //! #     "0.0.0.0:1234", // to address
 //! #     "namespace", // namespace for all metrics
 //! # );
@@ -63,8 +63,8 @@
 //! whatever you want, as long as it implements `AsRef<str>`.
 //!
 //! ```
-//! # use prima_datadog::{*, configuration::Config};
-//! # let configuration = Config::new(
+//! # use prima_datadog::{*, configuration::Configuration};
+//! # let configuration = Configuration::new(
 //! #     "0.0.0.0:1234", // to address
 //! #     "namespace", // namespace for all metrics
 //! # );
@@ -113,13 +113,13 @@
 
 use std::future::Future;
 
+use configuration::Configuration;
 pub use dogstatsd::{ServiceCheckOptions, ServiceStatus};
 use once_cell::sync::OnceCell;
 
 pub use client::DogstatsdClient;
 pub use tracker::*;
 
-use crate::configuration::Configuration;
 use crate::error::Error;
 
 mod client;
@@ -173,7 +173,7 @@ pub struct Datadog<C: DogstatsdClient> {
 impl Datadog<dogstatsd::Client> {
     /// Initializes a Datadog instance with a struct that implements the [Configuration] trait.
     /// Make sure that you run it only once otherwise you will get an error.
-    pub fn init(mut configuration: impl Configuration) -> Result<(), Error> {
+    pub fn init(mut configuration: Configuration) -> Result<(), Error> {
         let mut initialized: bool = false;
 
         // the closure is guaranteed to execute only once
